backend/requirements.txt
# ==============================================================================
# HL-OS Backend Python Dependencies
# ==============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# AI SDK
google-generativeai==0.3.2
anthropic==0.18.1

# HTTP Client
httpx==0.26.0

# Image Processing
Pillow==10.1.0

# Markdown & Frontmatter
python-frontmatter==1.0.1
python-slugify==8.0.1

# Redis
redis==5.0.1

# Database (optional, for future use)
sqlalchemy==2.0.25
alembic==1.13.1

# Logging
structlog==24.1.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
ruff==0.1.14
mypy==1.8.0

# CORS
starlette==0.35.1
